+++ Meraki.Api/Data/OrganizationAssuranceAlertNetwork.cs
@@ -4,4 +4,12 @@ namespace Meraki.Api.Data;
 /// Organization Assurance Alert Network
 /// </summary>
 [DataContract]
-public class OrganizationAssuranceAlertNetwork : NamedIdentifiedItem;
+public class OrganizationAssuranceAlertNetwork : NamedIdentifiedItem
+{
+	/// <summary>
+	/// Dashboard URL for the network.
+	/// </summary>
+	[ApiAccess(ApiAccess.Read)]
+	[DataMember(Name = "url")]
+	public string Url { get; set; } = string.Empty;
+}